chore(dependabot): add Dependabot configuration file to enable automated dependency updates
The `.github/dependabot.yml` file is added to the repository. This file contains the configuration for Dependabot, a tool that automates dependency updates.

The configuration specifies two update types:
1. `github-actions`: This update type is for updating dependencies related to GitHub Actions. The commit messages for these updates will have the prefix `ci(Github Action)`. The updates will be checked and applied daily.
2. `gitsubmodule`: This update type is for updating dependencies related to Git submodules. The updates will be checked and applied daily.

The `open-pull-requests-limit` is set to 1000, which means that Dependabot will create a maximum of 1000 open pull requests for updates.

This configuration will help ensure that the repository's dependencies are kept up to date automatically.