update article(19): add instructions for obtaining IPv6 PD prefix and automatically reacquiring it when PPPoE connection is disconnected
The commit adds instructions for obtaining the IPv6 PD prefix and adding it to the "lan" connection. It also includes a script to automatically disable and enable the PPPoE connection to reacquire the PD prefix when the connection is disconnected. This helps to ensure that devices in the local network can continue to use IPv6 even if the PPPoE connection is interrupted.