update article(5): add support for `host.txt` file in the L4D2 server container
The change adds a new volume mount in the services section of the `index.md` file. The `host.txt` file is now mounted to the `/home/steam/l4d2server/left4dead2/host.txt` path inside the container. This allows the server to access and use the `host.txt` file.